pastas e nomes: criando arquitetura e arrumando nomes
